pyproject.toml
[project]
name = "taxxa-ml-assessment"
version = "0.1.0"
description = "RAG Evaluation & Accounting Model Fine-Tuning — Taxxa ML Engineer Assessment"
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "openai>=1.30.0",

    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "ragas>=0.1.0",
    "datasets>=2.18.0",
    "langchain-community>=0.0.38",
]

[project.optional-dependencies]
# Install with: uv sync --extra finetune
finetune = [
    # Install PyTorch separately with the correct CUDA version:
    #   uv pip install torch --index-url https://download.pytorch.org/whl/cu128
    "torch>=2.3.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "trl>=0.8.6",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.29.0",
    "matplotlib>=3.8.0",
]

[tool.uv]
dev-dependencies = []
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"